Change Path to accept constructor
