--- stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/file/IsWritable.java
@@ -1,55 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.nativeimpl.file;
-
-import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BBoolean;
-import org.ballerinalang.model.values.BStruct;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.natives.annotations.ReturnType;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-/**
- * Determines whether the user has write access to the specified file.
- *
- * @since 0.94.1
- */
-@BallerinaFunction(
-        orgName = "ballerina", packageName = "file",
-        functionName = "isWritable",
-        receiver = @Receiver(type = TypeKind.STRUCT, structType = "File", structPackage = "ballerina.file"),
-        returnType = {@ReturnType(type = TypeKind.BOOLEAN)},
-        isPublic = true
-)
-public class IsWritable extends BlockingNativeCallableUnit {
-
-    @Override
-    public void execute(Context context) {
-        BStruct fileStruct = (BStruct) context.getRefArgument(0);
-        Path filePath = Paths.get(fileStruct.getStringField(0));
-        boolean isWritable = Files.isWritable(filePath);
-        context.setReturnValues(new BBoolean(isWritable));
-    }
-}
