+++ tests/ballerina-test/src/test/resources/test-src/taintchecking/connectors/file-read-negative.bal
@@ -2,8 +2,9 @@ import ballerina/io;
 import ballerina/file;
 
 function main (string[] args) {
-    file:Path target = file:getPath("/tmp/result.txt");
-    io:ByteChannel bchannel = check file:newByteChannel(target,args[0]);
+    file:Path target = new("/tmp/result.txt");
+    string absolutePath = target.toAbsolutePath().getPathValue();
+    io:ByteChannel bchannel = io:openFile(absolutePath, args[0]);
     int intArg = check <int> args[0];
     
     var readOutput = bchannel.read(intArg);
